API Reference

Functions

SphericalHarmonicTransforms.synthesize — Function

mapbuf = synthesize(ringpix::AbstractRingPixelization, alms::AbstractMatrix{<:Complex})

Performs the spherical harmonic synthesis transform of the harmonic coefficients alms for a real-valued map mapbuf, pixelized as described by ringpix.

See also synthesize!, analyze, analyze!

SphericalHarmonicTransforms.synthesize! — Function

synthesize!(mapbuf::AbstractArray{<:Real}, ringpix::AbstractRingPixelization, alms::AbstractMatrix{<:Complex})

Performs the spherical harmonic synthesis transform of the harmonic coefficients alms for a real-valued map mapbuf, pixelized as described by ringpix.

See also synthesize, analyze, analyze!

Extended help

The synthesis transform is implemented such that a real-valued function $f(θ, ϕ)$ is generated from complex-valued spherical harmonic coefficients $a_{ℓm}$ as

\[ f(θ, ϕ) = \sum_{ℓ=0}^{ℓ_\mathrm{max}} \sum_{m=0}^{ℓ} a_{ℓm} Y_{ℓm}(θ, ϕ)\]

where the $a_{ℓm}$s matrix contains only the non-negative orders ($m ≥ 0$) since the negative orders are not independent when $f(θ, ϕ)$ is real.

SphericalHarmonicTransforms.analyze — Function

alms = analyze(ringpix::AbstractRingPixelization, mapbuf::AbstractArray{<:Real}, lmax::Integer, mmax::Integer = lmax)

Performs the spherical harmonic analysis transform of the real-valued map mapbuf, pixelized as described by ringpix, to the corresponding complex-valued spherical harmonic coefficients alms.

See also analyze!, synthesize, synthesize!

SphericalHarmonicTransforms.analyze! — Function

analyze!(alms::AbstractMatrix{<:Complex}, ringpix::AbstractRingPixelization, mapbuf::AbstractAray{<:Real})

Performs the discrete spherical harmonic analysis transform of the real-valued map mapbuf, pixelized as described by ringpix, to the corresponding complex-valued spherical harmonic coefficients alms.

See also analyze, synthesize, synthesize!

Extended help

The analysis transform is implemented such that a real-valued function $f(θ, ϕ)$ is analyzed for its corresponding complex-valued spherical harmonic coefficients $a_{ℓm}$ as

\[ a_{ℓm} = \sum_{ℓ=0}^{ℓ_\mathrm{max}} \sum_{m=0}^ℓ f(θ, ϕ) \overline{Y_{ℓm}}(θ, ϕ) ΔΩ\]

where the overline indicates complex conjugation, $ΔΩ$ is the integration element (pixel area) on the sphere, and the $a_{ℓm}$s matrix contains only the non-negative orders ($m ≥ 0$) since the negative orders are not independent when $f(θ, ϕ)$ is real.

Ring-based Pixelizations

Pixelizations

SphericalHarmonicTransforms.RingPixelization — Type

RingPixelization{R} <: AbstractRingPixelization{R}

A pixelization of the unit sphere described by an arbitrary vector of Rings.

Examples

julia> ringpix = RingPixelization(ECPPixelization(5, 10))
5-ring RingPixelization{Float64} with 50 pixels

SphericalHarmonicTransforms.ECPPixelization — Type

ECPPixelization{R} <: AbstractRingPixelization{R}

An Equidistant Cylindrical Projection pixelization in column-major order, with dimensions nθ × nϕ in the colatitude/azimuth directions, respectively.

See also RingPixelization

Examples

julia> ECPPixelization(250, 500)
250×500 ECPPixelization{Float64}

julia> ECPPixelization{Float32}(250, 500)
250×500 ECPPixelization{Float32}

Extended help

The described pixelization covers the entire sphere $[0, π] × [0, 2π]$ with uniformly-sized pixels in coordinate space of size $Δθ = π/n_θ$ by $Δϕ = 2π/n_ϕ$. The pixel centers take on coordinates

\[ (θ_j, ϕ_k) = \left( π \frac{2j + 1}{2n_θ}, 2π \frac{2k + 1}{2n_ϕ} \right)\]

for integers $j ∈ \{0, …, n_θ - 1\}$ and $k ∈ \{0, …, n_ϕ - 1\}$. On the sphere, the pixels cover a solid angle $ΔΩ_{jk} ≈ \sin(θ_j) Δθ Δϕ$, i.e. the pixels' physical size approaches zero towards the poles.

See also equidistant cylindrical projection

Abstract Pixelization Interface

The following functions and types are considered part of the public interface, though they are unlikely to be used unless you are implementing a new pixelization.

SphericalHarmonicTransforms.AbstractRingPixelization — Type

abstract type AbstractRingPixelization{R} end

Abstract supertype of all ring map descriptor types.

SphericalHarmonicTransforms.Ring — Type

Ring{R<:Real}

Parameters required to describe an isolatitude ring on the sphere compatible with spherical harmonic transforms.

Fields

offset::Tuple{Int, Int}: Offsets (1-indexed) to the first pixel of the isolatitude ring(s) within a map array. The first offset in the tuple corresponds to the colatitude $θ$ given by .cosθ, while the second offset is for the ring mirrored over the equator (colatitude $π - θ$). Either offset may be 0 to indicate the ring is not present (such as for the equator, but also partial sphere coverage such as transforms on bands).

stride::Int: Stride between pixels within the ring.

nϕ::Int: The number of equispaced pixels in the isolatitude ring.

cosθ::R: Cosine of the colatitude of the ring, i.e. $\cos(θ)$.

ϕ₀_π::R: Azimuth of the first pixel in the isolatitude ring $ϕ₀$ divided by $π$.

ΔΩ::R: Solid angle area of pixels in the ring, roughly $\sin(θ) × Δθ × Δϕ$.

SphericalHarmonicTransforms.rings — Function

rr = rings(ringpix::AbstractRingPixelization)

A vector (or iterable) rr of Rings described by ringpix.

SphericalHarmonicTransforms.nring — Function

nr = nring(ringpix::AbstractRingPixelization)

Number of isolatitude rings nr described by ringpix.

SphericalHarmonicTransforms.npix — Function

np = npix(ringpix::AbstractRingPixelization)

Number of pixels np covered by all rings described by ringpix.

SphericalHarmonicTransforms.nϕmax — Function

n = nϕmax(ringpix::AbstractRingPixelization)

Number of pixels n in the longest isolatitude ring described by ringpix.

SphericalHarmonicTransforms.buffer — Function

mapbuf = buffer(ringpix::AbstractRingPixelization, ::Type{T})

Allocates an array mapbuf with element type T appropriate for containing a map in the pixelization scheme described by ringpix.